requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
deep-translator==1.11.4
deepl-translate==1.2.0
dill==0.3.6
easyocr==1.6.2
filelock==3.12.2
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
lazy_loader==0.3
loguru==0.7.0
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
multiprocess==0.70.14
networkx==3.1
ninja==1.11.1
numpy==1.25.1
opencv-contrib-python==4.5.5.62
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
packaging==23.1
pathos==0.3.0
Pillow==9.3.0
pox==0.3.2
ppft==1.7.6.6
pyclipper==1.3.0.post4
pycparser==2.21
PyExecJS==1.5.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
python-bidi==0.4.2
python-dotenv==1.0.0
PyWavelets==1.4.1
PyYAML==6.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
soupsieve==2.4.1
sympy==1.12
tifffile==2023.7.4
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
translators==5.7.9
typing_extensions==4.7.1
urllib3==2.0.3
win32-setctime==1.1.0
